backend/requirement.txt
fastapi
uvicorn[standard]
crewai
langchain-litellm
python-dotenv